requirements.txt
requests
python-dotenv
tqdm
pyinstaller